ui/requirements.txt
boto3
Pillow
pymupdf
Spire.Doc
pymupdf4llm
paddlepaddle 
paddleocr
tabulate
setuptools

# for local testing
streamlit